use long for max-arg-count
Use a signed type (as with argc) to avoid casting/sign conversions.

+++ src/firejail/main.c
@@ -1079,8 +1079,8 @@ int main(int argc, char **argv, char **envp) {
 	if (argc == 0 || !argv || strlen(argv[0]) == 0) {
 		fprintf(stderr, "Error: argv is invalid\n");
 		exit(1);
-	} else if ((unsigned long)argc >= max_arg_count) {
-		fprintf(stderr, "Error: too many arguments: argc (%d) >= max-arg-count (%lu)\n",
+	} else if (argc >= max_arg_count) {
+		fprintf(stderr, "Error: too many arguments: argc (%d) >= max-arg-count (%ld)\n",
 		        argc, max_arg_count);
 		exit(1);
 	}
@@ -1274,7 +1274,7 @@ int main(int argc, char **argv, char **envp) {
 			EUID_USER();}
 #endif
 
-			unsigned long j;
+			int j;
 			for (i = 1, j = fullargc; i < argc && j < max_arg_count; i++, j++, fullargc++)
 				fullargv[j] = argv[i];
